api/requirements.txt
fastapi
uvicorn
pydantic
google-genai
sse-starlette
python-dotenv
Pillow
